Fix 32-bit problems

+++ link-grammar/disjunct-utils.c
@@ -1177,11 +1177,11 @@ Tracon_sharing *pack_sentence_for_pruning(Sentence sent, unsigned int dcnt,
 	else
 	{
 		lgdebug(D_DISJ, "Debug: Encode for pruning (len %zu): "
-		        "tracon_id %zu (%zu+,%zu-), shared connectors %ld\n",
+		        "tracon_id %zu (%zu+,%zu-), shared connectors %d\n",
 		        sent->length,
 		        ts->tracon_list->entries[0]+ts->tracon_list->entries[1],
 		        ts->tracon_list->entries[0], ts->tracon_list->entries[1],
-				  &ts->cblock_base[ccnt_before] - ts->cblock);
+				  (int)(&ts->cblock_base[ccnt_before] - ts->cblock));
 	}
 
 	return ts;
@@ -1217,11 +1217,11 @@ Tracon_sharing *pack_sentence_for_parsing(Sentence sent, unsigned int dcnt,
 	else
 	{
 		lgdebug(D_DISJ, "Debug: Encode for parsing (len %zu): "
-		        "tracon_id %d (%d+,%d-), shared connectors %ld\n",
+		        "tracon_id %d (%d+,%d-), shared connectors %d\n",
 		        sent->length,
 		        (ts->next_id[0]-ts->word_offset)+(ts->next_id[1]-ts->word_offset),
 		        ts->next_id[0]-ts->word_offset, ts->next_id[1]-ts->word_offset,
-		        &ts->cblock_base[ccnt_before] - ts->cblock);
+		        (int)(&ts->cblock_base[ccnt_before] - ts->cblock));
 	}
 
 	return ts;
